refactor(feature/Layers/lib): создала отдельную папку под вспомогательный функции и обновила imports
